+++ H/ugens7.h
@@ -28,10 +28,10 @@
 
 typedef struct ovrlap {
         struct ovrlap  *nxtact, *nxtfree;
-        long    timrem, dectim, formphs, forminc, risphs, risinc, decphs, decinc;
+        int32    timrem, dectim, formphs, forminc, risphs, risinc, decphs, decinc;
         MYFLT   curamp, expamp;
         MYFLT   glissbas;/* Gliss factor to add to forminc (ifna index incr) */
-        long    sampct;         /* Sample count since grain started */
+        int32    sampct;         /* Sample count since grain started */
 } OVRLAP;
 
 typedef struct {
@@ -40,7 +40,7 @@ typedef struct {
         MYFLT   *iolaps, *ifna, *ifnb, *itotdur, *iphs, *ifmode, *iskip;
   /* kgliss and ifmode are same field */
         OVRLAP  basovrlap;
-        long    durtogo, fundphs, fofcount, prvsmps;
+        int32    durtogo, fundphs, fofcount, prvsmps;
         MYFLT   prvband, expamp, preamp;
         short   foftype;        /* Distinguish fof and fof2 */
         short   xincod, ampcod, fundcod, formcod, fmtmod;
@@ -52,14 +52,14 @@ typedef struct {
         OPDS    h;
         MYFLT   *ar, *asig, *kest, *kvar, *kfrq1, *kfrq2;
         MYFLT   *icpsmode, *ilowest, *iptrkprd;
-        long    nbufsmps, n2bufsmps, phase1, phase2, period, autoktim, autokcnt;
-        long    mindist, maxdist, max2dist, lomaxdist, cpsmode;
+        int32   nbufsmps, n2bufsmps, phase1, phase2, period, autoktim, autokcnt;
+        int32   mindist, maxdist, max2dist, lomaxdist, cpsmode;
         MYFLT   c1, c2, prvq, prvest, prvar, minfrq, estprd, lsicvt;
         MYFLT   *bufp, *midp, *inp1, *inp2;
         MYFLT   *bufq, *midq, *inq1, *inq2, *autobuf;
         MYFLT   *puls1, *puls2, *puls3, lin1, lin2, lin3;
         MYFLT   inc1, inc2, inc11, inc12, inc21, inc22, inc31, inc32;
-        long    cnt1, cnt2, cnt3, pnt1, pnt2, pnt3;
-        long    pnt11, pnt12, pnt13, pnt21, pnt22, pnt23, pnt31, pnt32, pnt33;
+        int32   cnt1, cnt2, cnt3, pnt1, pnt2, pnt3;
+        int32   pnt11, pnt12, pnt13, pnt21, pnt22, pnt23, pnt31, pnt32, pnt33;
         AUXCH   auxch;
 } HARMON;
